@@ -109,8 +109,6 @@ class Spark(Spark2):
         ]
 
     class Parser(Spark2.Parser):
-        OPTIONAL_ALIAS_TOKEN_CTE = True
-
         FUNCTIONS = {
             **Spark2.Parser.FUNCTIONS,
             "ANY_VALUE": _build_with_ignore_nulls(exp.AnyValue),
@@ -153,7 +151,7 @@ class Spark(Spark2):
             **Spark2.Generator.TYPE_MAPPING,
             exp.DataType.Type.MONEY: "DECIMAL(15, 4)",
             exp.DataType.Type.SMALLMONEY: "DECIMAL(6, 4)",
-            exp.DataType.Type.UNIQUEIDENTIFIER: "STRING",
+            exp.DataType.Type.UUID: "STRING",
             exp.DataType.Type.TIMESTAMPLTZ: "TIMESTAMP_LTZ",
             exp.DataType.Type.TIMESTAMPNTZ: "TIMESTAMP_NTZ",
         }
