@@ -1,7 +1,8 @@
 from __future__ import annotations
 
 from sqlglot import exp, generator, parser, tokens, transforms
-from sqlglot.dialects.dialect import Dialect, rename_func
+from sqlglot.dialects.dialect import Dialect, rename_func, strposition_sql as _strposition_sql
+from sqlglot.helper import seq_get
 
 
 class Tableau(Dialect):
@@ -39,9 +40,22 @@ class Tableau(Dialect):
                 return self.func("COUNTD", *this.expressions)
             return self.func("COUNT", this)
 
+        def strposition_sql(self, expression: exp.StrPosition) -> str:
+            has_occurrence = "occurrence" in expression.args
+            return _strposition_sql(
+                self,
+                expression,
+                func_name="FINDNTH" if has_occurrence else "FIND",
+                supports_occurrence=has_occurrence,
+            )
+
     class Parser(parser.Parser):
         FUNCTIONS = {
             **parser.Parser.FUNCTIONS,
             "COUNTD": lambda args: exp.Count(this=exp.Distinct(expressions=args)),
+            "FIND": exp.StrPosition.from_arg_list,
+            "FINDNTH": lambda args: exp.StrPosition(
+                this=seq_get(args, 0), substr=seq_get(args, 1), occurrence=seq_get(args, 2)
+            ),
         }
         NO_PAREN_IF_COMMANDS = False
