@@ -118,7 +118,7 @@ def simplify(
         node = flatten(node)
         node = simplify_connectors(node, root)
         node = remove_complements(node, root)
-        node = simplify_coalesce(node)
+        node = simplify_coalesce(node, dialect)
         node.parent = expression.parent
         node = simplify_literals(node, root)
         node = simplify_equality(node)
@@ -750,7 +750,7 @@ def simplify_parens(expression):
 
     if (
         not isinstance(this, exp.Select)
-        and not isinstance(parent, exp.SubqueryPredicate)
+        and not isinstance(parent, (exp.SubqueryPredicate, exp.Bracket))
         and (
             not isinstance(parent, (exp.Condition, exp.Binary))
             or isinstance(parent, exp.Paren)
@@ -776,7 +776,7 @@ def _is_constant(expression: exp.Expression) -> bool:
     return isinstance(expression, exp.CONSTANTS) or _is_date_literal(expression)
 
 
-def simplify_coalesce(expression):
+def simplify_coalesce(expression: exp.Expression, dialect: DialectType) -> exp.Expression:
     # COALESCE(x) -> x
     if (
         isinstance(expression, exp.Coalesce)
@@ -785,6 +785,12 @@ def simplify_coalesce(expression):
         and not isinstance(expression.parent, exp.Hint)
     ):
         return expression.this
+
+    # We can't convert `COALESCE(x, 1) = 2` into `NOT x IS NULL AND x = 2` for redshift,
+    # because they are not always equivalent. For example,  if `x` is `NULL` and it comes
+    # from a table, then the result is `NULL`, despite `FALSE AND NULL` evaluating to `FALSE`
+    if dialect == "redshift":
+        return expression
 
     if not isinstance(expression, COMPARISONS):
         return expression
