@@ -616,10 +616,10 @@ class TestExecutor(unittest.TestCase):
             ("CONCAT('a', 'b')", "ab"),
             ("CONCAT('a', NULL)", None),
             ("CONCAT_WS('_', 'a', 'b')", "a_b"),
-            ("STR_POSITION('bar', 'foobarbar')", 4),
-            ("STR_POSITION('bar', 'foobarbar', 5)", 7),
-            ("STR_POSITION(NULL, 'foobarbar')", None),
-            ("STR_POSITION('bar', NULL)", None),
+            ("STR_POSITION('foobarbar', 'bar')", 4),
+            ("STR_POSITION('foobarbar', 'bar', 5)", 7),
+            ("STR_POSITION('foobarbar', NULL)", None),
+            ("STR_POSITION(NULL, 'bar')", None),
             ("UPPER('foo')", "FOO"),
             ("UPPER(NULL)", None),
             ("LOWER('FOO')", "foo"),
