@@ -1,6 +1,6 @@
-import sys
 import datetime
 import math
+import sys
 import unittest
 
 from sqlglot import ParseError, alias, exp, parse_one
@@ -55,7 +55,6 @@ class TestExpressions(unittest.TestCase):
             parse_one("ROW() OVER(Partition by y)"),
             parse_one("ROW() OVER (partition BY y)"),
         )
-        self.assertEqual(parse_one("TO_DATE(x)", read="hive"), parse_one("ts_or_ds_to_date(x)"))
         self.assertEqual(exp.Table(pivots=[]), exp.Table())
         self.assertNotEqual(exp.Table(pivots=[None]), exp.Table())
         self.assertEqual(
@@ -275,6 +274,16 @@ class TestExpressions(unittest.TestCase):
                 {
                     "`a-b`.`c`": parse_one("select 1"),
                 },
+                dialect="spark",
+            ).sql(),
+            "SELECT * FROM (SELECT 1) AS a /* source: a-b.c */",
+        )
+
+    def test_expand_with_lazy_source_provider(self):
+        self.assertEqual(
+            exp.expand(
+                parse_one('select * from "a-b"."C" AS a'),
+                {"`a-b`.c": lambda: parse_one("select 1", dialect="spark")},
                 dialect="spark",
             ).sql(),
             "SELECT * FROM (SELECT 1) AS a /* source: a-b.c */",
@@ -637,6 +646,12 @@ class TestExpressions(unittest.TestCase):
         self.assertTrue(all(isinstance(e, exp.Expression) for e in expression.walk()))
         self.assertTrue(all(isinstance(e, exp.Expression) for e in expression.walk(bfs=False)))
 
+    def test_str_position_order(self):
+        str_position_exp = parse_one("STR_POSITION('mytest', 'test')")
+        self.assertIsInstance(str_position_exp, exp.StrPosition)
+        self.assertEqual(str_position_exp.args.get("this").this, "mytest")
+        self.assertEqual(str_position_exp.args.get("substr").this, "test")
+
     def test_functions(self):
         self.assertIsInstance(parse_one("x LIKE ANY (y)"), exp.Like)
         self.assertIsInstance(parse_one("x ILIKE ANY (y)"), exp.ILike)
@@ -841,6 +856,7 @@ class TestExpressions(unittest.TestCase):
 
     def test_convert(self):
         from collections import namedtuple
+
         import pytz
 
         PointTuple = namedtuple("Point", ["x", "y"])
