@@ -40,7 +40,7 @@ class TestJsonpath(unittest.TestCase):
                 self.assertEqual(parse(selector).sql(), f"'{expected}'")
 
     def test_cts_file(self):
-        with open(os.path.join(FIXTURES_DIR, "jsonpath", "cts.json")) as file:
+        with open(os.path.join(FIXTURES_DIR, "jsonpath", "cts.json"), encoding="utf-8") as file:
             tests = json.load(file)["tests"]
 
         # sqlglot json path generator rewrites to a normal form
